# app/internal/app/api/v1/openapi/types.gen.go
// Package openapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.9.0 DO NOT EDIT.
package openapi

const (
	BearerScopes = "Bearer.Scopes"
)

// Team defines model for Team.
type Team struct {
	Description string `json:"description"`
	Id          string `json:"id"`
	Name        string `json:"name"`
	Users       []User `json:"users"`
}

// User defines model for User.
type User struct {
	Id   string `json:"id"`
	Name string `json:"name"`
}

// PutProfileJSONBody defines parameters for PutProfile.
type PutProfileJSONBody struct {
	Name *string `json:"name,omitempty"`
}

// GetTeamsParams defines parameters for GetTeams.
type GetTeamsParams struct {
	// ページ数
	Page int `json:"page"`

	// 1ページあたりの取得数
	Limit int `json:"limit"`

	// 名前を接頭辞で検索
	Name *string `json:"name,omitempty"`
}

// PostTeamsJSONBody defines parameters for PostTeams.
type PostTeamsJSONBody struct {
	Description string `json:"description"`
	Name        string `json:"name"`
}

// PutTeamsTeamIdJSONBody defines parameters for PutTeamsTeamId.
type PutTeamsTeamIdJSONBody struct {
	Description *string `json:"description,omitempty"`
	Name        *string `json:"name,omitempty"`
}

// PostUserJSONBody defines parameters for PostUser.
type PostUserJSONBody struct {
	Name string `json:"name"`
}

// PutProfileJSONRequestBody defines body for PutProfile for application/json ContentType.
type PutProfileJSONRequestBody PutProfileJSONBody

// PostTeamsJSONRequestBody defines body for PostTeams for application/json ContentType.
type PostTeamsJSONRequestBody PostTeamsJSONBody

// PutTeamsTeamIdJSONRequestBody defines body for PutTeamsTeamId for application/json ContentType.
type PutTeamsTeamIdJSONRequestBody PutTeamsTeamIdJSONBody

// PostUserJSONRequestBody defines body for PostUser for application/json ContentType.
type PostUserJSONRequestBody PostUserJSONBody