# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: sample api
  version: "1.0"
  description: "ユーザーとチームが存在し、ユーザーはチームに参加することができる"
servers:
  - url: "http://localhost:8080/api/v1"
tags:
  - name: self
    description: リクエストを発行したユーザーに関する操作を行うもの
  - name: user
  - name: team

paths:
  /profile:
    get:
      summary: プロフィール取得
      tags:
        - self
      operationId: get-profile
      security:
        - Bearer: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    put:
      summary: プロフィール編集
      tags:
        - self
      operationId: put-profile
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
      responses:
        "204":
          description: No Content

  /users:
    post:
      summary: ユーザー作成
      tags:
        - user
      operationId: post-user
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required:
                - name
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /teams:
    get:
      summary: チーム一覧取得
      tags:
        - team
      operationId: get-teams
      security:
        - Bearer: []
      parameters:
        - schema:
            type: integer
          in: query
          name: page
          description: ページ数
          required: true
        - schema:
            type: integer
          in: query
          name: limit
          description: 1ページあたりの取得数
          required: true
        - schema:
            type: string
          in: query
          name: name
          description: 名前を接頭辞で検索
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"

    post:
      summary: チーム作成
      tags:
        - team
      operationId: post-teams
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
              required:
                - name
                - description
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"

  "/teams/{teamId}":
    parameters:
      - schema:
          type: string
        name: teamId
        in: path
        required: true

    put:
      summary: チーム編集
      tags:
        - team
      operationId: put-teams-teamId
      security:
        - Bearer: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "204":
          description: No Content

    delete:
      summary: チーム削除
      tags:
        - team
      operationId: delete-teams-teamId
      security:
        - Bearer: []
      responses:
        "204":
          description: No Content

  "/teams/{teamId}/join":
    parameters:
      - schema:
          type: string
        name: teamId
        in: path
        required: true

    post:
      summary: チーム参加
      description: すでに参加済みの場合はエラーを返す。
      tags:
        - self
        - team
      operationId: post-teams-teamId-join
      security:
        - Bearer: []
      responses:
        "201":
          description: Created

components:
  schemas:
    User:
      title: User
      type: object
      properties:
        id:
          type: string
          example: user_id
        name:
          type: string
          example: user_name
      required:
        - id
        - name

    Team:
      title: Team
      type: object
      properties:
        id:
          type: string
          example: team_id
        name:
          type: string
          example: team_name
        description:
          type: string
          example: team_description
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
      required:
        - id
        - name
        - description
        - users

  securitySchemes:
    Bearer:
      type: http
      scheme: bearer